consolidate submitted-lead files into lib/submitted-leads
