+++ apps/web/lib/submitted-leads/get-submitted-lead-or-throw.ts
@@ -1,5 +1,5 @@
 import { prisma } from "@dub/prisma";
-import { DubApiError } from "../errors";
+import { DubApiError } from "../api/errors";
 
 interface GetSubmittedLeadOrThrowParams {
   leadId: string;
